requirements.txt
asgiref==3.6.0
cx-Oracle==8.3.0
Django==4.0.3
djangorestframework==3.14.0
pytz==2021.3
sqlparse==0.4.2
Pillow==9.5.0